feat: start on transcoder
well this is a mess, I tried using an FFMPEG wrapper, but it crashed when I wanted to output a stream rather than writing content to disk.

turns out most of the ffmpeg wrappers that exist are out of date, so I'm going to write wrappers for all the individual encoding frontends.